# refactor(core): serialize response once, closes #5641