--- a/core/tauri/src/test/mod.rs
+++ b/core/tauri/src/test/mod.rs
@@ -58,7 +58,7 @@ use serialize_to_javascript::DefaultTemplate;
 use std::{borrow::Cow, collections::HashMap, fmt::Debug};
 
 use crate::{
-  ipc::{InvokeBody, InvokeError, InvokeResponse, RuntimeAuthority},
+  ipc::{InvokeError, InvokeResponse, InvokeResponseBody, RuntimeAuthority},
   webview::InvokeRequest,
   App, Assets, Builder, Context, Pattern, Runtime, Webview,
 };
@@ -282,7 +282,7 @@ pub fn assert_ipc_response<
 pub fn get_ipc_response<W: AsRef<Webview<MockRuntime>>>(
   webview: &W,
   request: InvokeRequest,
-) -> Result<InvokeBody, serde_json::Value> {
+) -> Result<InvokeResponseBody, serde_json::Value> {
   let (tx, rx) = std::sync::mpsc::sync_channel(1);
   webview.as_ref().clone().on_message(
     request,