… core/tauri-config-schema/schema.json
… core/tauri/src/ipc/format_callback.rs
… tooling/cli/schema.json